+++ example/android/app/src/main/kotlin/com/example/fluxy_example/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.example.fluxy_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity : FlutterActivity()